move base Analysis to its own module, related to #31; test decorators
